+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 ## Release 5.0.M11
 
 * #604 Allow to exclude null properties in responses
+* #605 Hide deprecated protocol keys from OpenAPI descriptors
 
 ## Release 5.0.M10